+++ src/main/scala/com/ogunleye/tenii/payments/model/db/TellerTeniiPot.scala
@@ -2,4 +2,4 @@ package com.ogunleye.tenii.payments.model.db
 
 import org.bson.types.ObjectId
 
-case class TellerTeniiPot(id: Option[ObjectId] = None, tellerUserId: String, limit: Double, amount: Double)
+case class TellerTeniiPot(id: Option[ObjectId] = None, tellerUserId: String, limit: Int, amount: Double)
